# Frontend/swagger-openapi.json
{
    "openapi": "3.0.0",
...
    "info": {
        "title": "Identity API",
        "description": "An ASP.NET Core Web API for managing identity items",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Contact",
            "url": "https://example.com/contact"
        },
        "license": {
            "name": "License",
            "url": "https://example.com/license"
        },
        "version": "v1"
    },
    "paths": {
...
        "/api/User/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "GetUserById",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTOApiResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTOApiResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDTOApiResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "Bearer": []
        }
    ],
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\n"
            }
        },
        "schemas": {
            "AssigningService": {
                "type": "object",
                "properties": {
                    "guid": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "key": {
                        "type": "string"
                    },
                    "dateActive": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "dateExpired": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
...
            "RoleResponse": {
                "type": "object",
                "properties": {
                    "guid": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "value": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "additionalProperties": false
            },
...
            "UserDTO": {
                "type": "object",
                "properties": {
                    "userId": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "avatarUrl": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "phoneCode": {
                        "type": "string"
                    },
                    "region": {
                        "type": "string"
                    },
                    "guid": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "dateCreated": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "dateModified": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "middleName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "isRequiredChangePassword": {
                        "type": "boolean"
                    },
                    "isPasswordExpired": {
                        "type": "boolean"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "isLogin": {
                        "type": "boolean"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "otpCode": {
                        "type": "string"
                    },
                    "attemptLogin": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "locked": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "lastLogin": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "expirationDate": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "otpLifeTime": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "step": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AssigningService"
                        }
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleResponse"
                        }
                    }
                },
                "additionalProperties": false
            },
            "UserDTOApiResponse": {
                "type": "object",
                "properties": {
                    "requestId": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "isSuccess": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserDTO"
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}